newInstance function is moved from abstract class to the right class.
