+++ app/src/main/java/org/gdg/frisbee/android/eventseries/GdgEventListFragment.java
@@ -1,5 +1,7 @@
 package org.gdg.frisbee.android.eventseries;
 
+import android.os.Bundle;
+
 import com.android.volley.Response;
 
 import org.gdg.frisbee.android.Const;
@@ -17,6 +19,15 @@ import de.keyboardsurfer.android.widget.crouton.Crouton;
 import de.keyboardsurfer.android.widget.crouton.Style;
 
 public class GdgEventListFragment extends EventListFragment {
+
+    public static EventListFragment newInstance(String plusId) {
+        EventListFragment fragment = new GdgEventListFragment();
+        Bundle arguments = new Bundle();
+        arguments.putString(Const.EXTRA_PLUS_ID, plusId);
+        fragment.setArguments(arguments);
+        return fragment;
+    }
+    
     @Override
     void fetchEvents() {
         final DateTime now = new DateTime();
@@ -65,5 +76,4 @@ public class GdgEventListFragment extends EventListFragment {
             });
         }
     }
-
 }
